Add basic selenium test
